backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
sqlalchemy==2.0.25
greenlet==3.0.3
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
passlib[bcrypt]==1.7.4
# Passlib 1.7.x expects bcrypt to expose __about__.__version__.
# Pinning to 3.2.2 avoids noisy runtime warnings during startup.
bcrypt==3.2.2
PyJWT==2.8.0
pymc==5.16.1
numpy==1.26.4
scipy==1.11.4
pytest==7.4.4
httpx==0.26.0
python-multipart==0.0.9
